requirements.txt
riotwatcher
discord.py